cleanup: fix (harmless) use before assign
While looking around for possible wrong CNS values in the logbook header,
I see this "interesting" thing. Simply assign the current CNS data to the
global state, before using it to issue a warning. In practice, this does
not fix any real bug, as the code is run very often, so the global state
data is always (very) recent. It simply does not look right.

Signed-off-by: Jan Mulder <jlmulder@xs4all.nl>