New feature: make selected customviews persistent
Add the set of selected customviews during dive time to the
persistent settings. Nothing has changed to the menu, so a user
still has to access the on/off page from the simulator or during
an actual dive, but once changed here, the settings are preserved
for later dives. It feels a bit as waste of menu space to add a second
page for something that is already available. And further, I have to
admit the adding a page is rather complex to get the navigation right,
so that is a 2nd argument not to add a page.

Further, getting this one setting saved to the persist
preferences is a bit of a hack. Misused an existing global bool
value (WriteSettings) to force a actual write despite not being
in surface mode.

Signed-off-by: Jan Mulder <jlmulder@xs4all.nl>