New feature: do not show sensor customviews when no sensors attached
There are mCCR divers, that use a DC just for deco and time purposes
and not for O2 tracking and logging. These divers typically use the
existing Fixed SP mode, and have 0 active 02 sensors.

This commit simply suppresses display of the 2 customviews in case
there are no active sensors defined.

Signed-off-by: Jan Mulder <jlmulder@xs4all.nl>