Bugfix: write NDL in logbook correctly
This is a subtle one. Typecasting takes precedence over, for example, division.
So, we first typecasted an int to an uint8_t, loosing data in the process, and
then do a seconds to minute conversion. This, obviously, does not give the
result one expects.

And in hindsight a trivial fix for a bug that like has been there forever.

Signed-off-by: Jan Mulder <jlmulder@xs4all.nl>