bugfix, consistency: show deco/NDL really after 1 minute divetime
Related to commit 02d10d955be2. We intended to show deco/NDL after
1 minute divetime, but we included the surface seconds as well.
And yes, this can hardly be called a bugfix. Its simply more
consistent to relate the actual divetime on the display (which
is without surface seconds) with the appearance of deco/NDL after
exactly 1 minute, instead of about 1 minute.

Signed-off-by: Jan Mulder <jlmulder@xs4all.nl>