cleanup: factor out unused stopwatch code
Visually, this looks a bit of a dangerous cleanup, but its not :-)

To explain this cleanup: stopwatch_seconds is never used, so assign
to it is useless, and the variable can be removed. Now,
stopwatch_start_at_this_dive_time_seconds is not used any more, so
remove it as well. Now, look at the boolResetStopwatch code. It
is used but only to reset itself. This makes the whole stopwatch
block in data_exchange_main.c a useless thing. All the other
cleanup is trivial.

Signed-off-by: Jan Mulder <jlmulder@xs4all.nl>