Bugfix: deco planner did not initialize properly
Redoing a deco calculation in the planner, after a change in used gasses, did result
in wrong gas usage figures. And a restart after sleep made things right again. This all
feels like some initialization problem. Nothing very important as it's only a desktop
planner, but a bug it is.

Code reading, I found the not correctly initialized data.

Reported-by: Teun (forum user tdewilt2)
Signed-off-by: Jan Mulder <jan@jlmulder.nl>